+++ CHANGELOG.md
@@ -12,8 +12,8 @@
   - [Security] для обновлений безопасности.
 
 ## [ToDo]
-- Create monitoring inrastracture (prometheus + grafana)
 - Add infra node to monitoring
+- Add stage and prod nodes to monitoring
 - Add Gitlab to monitoring
 - Create Grafana Dashboards
 - Add supported version of software in Readme
@@ -23,6 +23,22 @@
 - [Terraform] Repace paths to start terraform from terraform folder
 
 ## [Unreleased]
+
+### [0.1.7] 2019-10-29
+### Fixed
+- [Ansible] Role: monitoring. Run handler only if docker stack not changed, but config is changed
+- [Ansible] Fix monitoring group_vars
+
+### Added
+- [Terraform] Add crawler engine metrics port to vars. Create firewall rules
+- [Ansible] Add handler for prometheus
+- [Docker] Add trickster image
+- [Docker] Add grafana image
+- [Ansible] Add grafana and trickster to monitoring role
+- [Ansible] Add alertmanager to monitoring role
+- [Docker] Add alertmanager image
+- [Terraform] Add firewall rule for metrics ports in stage and prod
+- [Ansible] Role: monitoring. Add node-exporter to compose file 
 
 ## [0.1.6] 2019-10-27
 ### Added